Hoisting
Wherever a var appears inside a scope, that declaration is taken to
belong to the entire scope and accessible everywhere throughout.
Metaphorically, this behavior is called hoisting, when a var declaration
is conceptually “moved” to the top of its enclosing scope.

If you try to access a variable’s value in a scope where it’s not avail
able, you’ll get a ReferenceError thrown.